UMI4C class methods

This page contains a summary of the different methods used to access the information contained inside the UMI4C-class object. See the details section for more information on the different accessors.

dgram(object)

dgram(object) <- value

bait(object)

trend(object)

results(object, format = "GRanges", counts = TRUE, ordered = FALSE)

# S4 method for UMI4C
dgram(object)

# S4 method for UMI4C
dgram(object) <- value

# S4 method for UMI4C
bait(object)

# S4 method for UMI4C
trend(object)

# S4 method for UMI4C
results(object, format = "GRanges", counts = FALSE, ordered = FALSE)

Arguments

object
a UMI4C-class object.

value
Alternative list of dgrams to replace the current slot.

format
Either "GRanges" (default) or "data.frame", indicating the format output of the results.

counts
Logical indicating whether counts for the different region should be provided. Default: FALSE.

ordered
Logical indicating whether to sort output by significance (adjusted p-value). Default: FALSE.

Value

There are several accessors to easily retrive information from a UMI4C-class object:

dgram: Returns a named list with the output domainograms for each sample.

bait: Returns a GRanges object with the position of the bait.

trend: Returns a data.frame in long format with the values of the adapative smoothen trend.

results: Returns a GRanges or data.frame with the results of the differential analysis.

See also

Examples

# Create example UMI4C-class object with helper function
umi <- makeUMI4Cexample()

#> Joining, by = "sampleID"

# Access the different information inside the UMI4C object
dgram(umi)

#> List of length 2
#> names(2): ctrl cyt

bait(umi)

#> GRanges object with 1 range and 1 metadata column:
#>                   seqnames    ranges strand |        name
#>                      <Rle> <IRanges>  <Rle> | <character>
#>   ctrl_hi19_SOCS1    chr16  11349721      * |       SOCS1
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

trend(umi)

#>     geo_coord     trend         sd scale id_contact sample
#> 1    10973745 0.3900000 0.07211103   150 frag_26766   ctrl
#> 2    10974076 0.3859060 0.07197185   149 frag_26767   ctrl
#> 3    10974410 0.3885135 0.07245814   148 frag_26768   ctrl
#> 4    10974743 0.3911565 0.07295106   147 frag_26769   ctrl
#> 5    10975075 0.3938356 0.07345072   146 frag_26770   ctrl
#> 6    10975404 0.3965517 0.07395728   145 frag_26771   ctrl
#> 7    10975734 0.3923611 0.07382046   144 frag_26772   ctrl
#> 8    10976066 0.3951049 0.07433668   143 frag_26773   ctrl
#> 9    10976398 0.3978873 0.07486018   142 frag_26774   ctrl
#> 10   10976730 0.4007092 0.07539111   141 frag_26775   ctrl
#> 11   10977058 0.3821429 0.07388629   140 frag_26776   ctrl
#> 12   10977439 0.3936170 0.07472095   141 frag_26777   ctrl
#> 13   10977793 0.3936170 0.07472095   141 frag_26778   ctrl
#> 14   10978120 0.3964286 0.07525467   140 frag_26779   ctrl
#> 15   10978474 0.3928571 0.07491492   140 frag_26780   ctrl
#> 16   10978852 0.3971631 0.07505677   141 frag_26781   ctrl
#> 17   10979204 0.4042553 0.07572396   141 frag_26782   ctrl
#> 18   10979532 0.4071429 0.07626484   140 frag_26783   ctrl
#> 19   10979862 0.4064748 0.07647587   139 frag_26784   ctrl
#> 20   10980195 0.4094203 0.07703004   138 frag_26785   ctrl
#> 21   10980528 0.4124088 0.07759231   137 frag_26786   ctrl
#> 22   10980857 0.4154412 0.07816284   136 frag_26787   ctrl
#> 23   10981189 0.4185185 0.07874182   135 frag_26788   ctrl
#> 24   10981523 0.4216418 0.07932945   134 frag_26789   ctrl
#> 25   10981880 0.4216418 0.07932945   134 frag_26790   ctrl
#> 26   10982259 0.4185185 0.07874182   135 frag_26791   ctrl
#> 27   10982592 0.4216418 0.07932945   134 frag_26792   ctrl
#> 28   10982948 0.4365672 0.08072130   134 frag_26793   ctrl
#> 29   10983280 0.4398496 0.08132822   133 frag_26794   ctrl
#> 30   10983616 0.4431818 0.08194435   132 frag_26795   ctrl
#> 31   10983950 0.4465649 0.08256988   131 frag_26796   ctrl
#> 32   10984284 0.4500000 0.08320503   130 frag_26797   ctrl
#> 33   10984620 0.4534884 0.08385003   129 frag_26798   ctrl
#> 34   10984958 0.4570312 0.08450511   128 frag_26799   ctrl
#> 35   10985293 0.4566929 0.08480574   127 frag_26800   ctrl
#> 36   10985630 0.4603175 0.08547881   126 frag_26801   ctrl
#> 37   10986011 0.4763780 0.08661417   127 frag_26802   ctrl
#> 38   10986374 0.4724409 0.08625552   127 frag_26803   ctrl
#> 39   10986714 0.4761905 0.08694009   126 frag_26804   ctrl
#> 40   10987058 0.4800000 0.08763561   125 frag_26805   ctrl
#> 41   10987424 0.4880000 0.08836289   125 frag_26806   ctrl
#> 42   10987772 0.4879032 0.08870968   124 frag_26807   ctrl
#> 43   10988122 0.4878049 0.08906058   123 frag_26808   ctrl
#> 44   10988491 0.4959350 0.08979968   123 frag_26809   ctrl
#> 45   10988843 0.5000000 0.09053575   122 frag_26810   ctrl
#> 46   10989189 0.4958678 0.09053265   121 frag_26811   ctrl
#> 47   10989535 0.5000000 0.09128709   120 frag_26812   ctrl
#> 48   10989875 0.5042017 0.09205421   119 frag_26813   ctrl
#> 49   10990216 0.5084746 0.09283433   118 frag_26814   ctrl
#> 50   10990559 0.5128205 0.09362779   117 frag_26815   ctrl
#> 51   10990905 0.5172414 0.09443492   116 frag_26816   ctrl
#> 52   10991253 0.5217391 0.09525610   115 frag_26817   ctrl
#> 53   10991598 0.5219298 0.09569046   114 frag_26818   ctrl
#> 54   10991946 0.5265487 0.09653728   113 frag_26819   ctrl
#> 55   10992298 0.5267857 0.09698911   112 frag_26820   ctrl
#> 56   10992645 0.5225225 0.09703000   111 frag_26821   ctrl
#> 57   10992991 0.5045455 0.09577867   110 frag_26822   ctrl
#> 58   10993332 0.5091743 0.09665737   109 frag_26823   ctrl
#> 59   10993722 0.5409091 0.09917011   110 frag_26824   ctrl
#> 60   10994066 0.5458716 0.10007993   109 frag_26825   ctrl
#> 61   10994414 0.5509259 0.10100659   108 frag_26826   ctrl
#> 62   10994765 0.5560748 0.10195058   107 frag_26827   ctrl
#> 63   10995120 0.5613208 0.10291238   106 frag_26828   ctrl
#> 64   10995471 0.5571429 0.10301575   105 frag_26829   ctrl
#> 65   10995825 0.5625000 0.10400629   104 frag_26830   ctrl
#> 66   10996179 0.5679612 0.10501606   103 frag_26831   ctrl
#> 67   10996524 0.5539216 0.10421712   102 frag_26832   ctrl
#> 68   10996872 0.5594059 0.10524897   101 frag_26833   ctrl
#> 69   10997217 0.5650000 0.10630146   100 frag_26834   ctrl
#> 70   10997567 0.5656566 0.10689904    99 frag_26835   ctrl
#> 71   10997918 0.5714286 0.10798985    98 frag_26836   ctrl
#> 72   10998273 0.5773196 0.10910315    97 frag_26837   ctrl
#> 73   10998631 0.5833333 0.11023964    96 frag_26838   ctrl
#> 74   10998988 0.5842105 0.11090162    95 frag_26839   ctrl
#> 75   10999345 0.5904255 0.11208142    94 frag_26840   ctrl
#> 76   10999706 0.5967742 0.11328660    93 frag_26841   ctrl
#> 77   11000108 0.6117021 0.11408304    94 frag_26842   ctrl
#> 78   11000465 0.6182796 0.11530973    93 frag_26843   ctrl
#> 79   11000842 0.6182796 0.11530973    93 frag_26844   ctrl
#> 80   11001235 0.6117021 0.11408304    94 frag_26845   ctrl
#> 81   11001625 0.6473684 0.11674249    95 frag_26846   ctrl
#> 82   11001982 0.6542553 0.11798443    94 frag_26847   ctrl
#> 83   11002341 0.6612903 0.11925308    93 frag_26848   ctrl
#> 84   11002702 0.6684783 0.12054931    92 frag_26849   ctrl
#> 85   11003066 0.6758242 0.12187403    91 frag_26850   ctrl
#> 86   11003424 0.6833333 0.12322818    90 frag_26851   ctrl
#> 87   11003786 0.6910112 0.12461277    89 frag_26852   ctrl
#> 88   11004151 0.6704545 0.12344069    88 frag_26853   ctrl
#> 89   11004518 0.6781609 0.12485955    87 frag_26854   ctrl
#> 90   11004881 0.6860465 0.12631140    86 frag_26855   ctrl
#> 91   11005236 0.6941176 0.12779742    85 frag_26856   ctrl
#> 92   11005597 0.7023810 0.12931882    84 frag_26857   ctrl
#> 93   11005963 0.7048193 0.13032113    83 frag_26858   ctrl
#> 94   11006334 0.7012195 0.13077811    82 frag_26859   ctrl
#> 95   11006706 0.7098765 0.13239266    81 frag_26860   ctrl
#> 96   11007068 0.7187500 0.13404757    80 frag_26861   ctrl
#> 97   11007435 0.7278481 0.13574437    79 frag_26862   ctrl
#> 98   11007799 0.7371795 0.13748468    78 frag_26863   ctrl
#> 99   11008155 0.7467532 0.13927020    77 frag_26864   ctrl
#> 100  11008512 0.7565789 0.14110270    76 frag_26865   ctrl
#>                             sampleID  replicate    viewpoint
#> 1   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 2   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 3   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 4   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 5   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 6   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 7   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 8   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 9   ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 10  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 11  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 12  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 13  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 14  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 15  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 16  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 17  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 18  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 19  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 20  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 21  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 22  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 23  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 24  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 25  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 26  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 27  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 28  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 29  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 30  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 31  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 32  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 33  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 34  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 35  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 36  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 37  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 38  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 39  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 40  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 41  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 42  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 43  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 44  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 45  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 46  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 47  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 48  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 49  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 50  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 51  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 52  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 53  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 54  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 55  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 56  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 57  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 58  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 59  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 60  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 61  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 62  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 63  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 64  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 65  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 66  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 67  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 68  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 69  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 70  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 71  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 72  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 73  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 74  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 75  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 76  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 77  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 78  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 79  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 80  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 81  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 82  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 83  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 84  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 85  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 86  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 87  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 88  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 89  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 90  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 91  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 92  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 93  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 94  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 95  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 96  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 97  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 98  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 99  ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#> 100 ctrl_hi19_SOCS1, ctrl_hi22_SOCS1 hi19, hi22 SOCS1, SOCS1
#>                                                                                                                                                                         file
#> 1   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 2   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 3   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 4   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 5   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 6   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 7   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 8   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 9   /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 10  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 11  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 12  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 13  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 14  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 15  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 16  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 17  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 18  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 19  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 20  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 21  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 22  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 23  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 24  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 25  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 26  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 27  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 28  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 29  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 30  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 31  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 32  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 33  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 34  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 35  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 36  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 37  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 38  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 39  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 40  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 41  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 42  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 43  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 44  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 45  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 46  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 47  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 48  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 49  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 50  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 51  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 52  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 53  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 54  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 55  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 56  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 57  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 58  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 59  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 60  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 61  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 62  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 63  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 64  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 65  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 66  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 67  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 68  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 69  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 70  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 71  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 72  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 73  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 74  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 75  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 76  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 77  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 78  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 79  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 80  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 81  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 82  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 83  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 84  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 85  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 86  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 87  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 88  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 89  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 90  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 91  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 92  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 93  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 94  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 95  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 96  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 97  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 98  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 99  /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#> 100 /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi19_SOCS1_counts.tsv.gz, /Users/mramos/tools/UMI4Cats/inst/extdata/SOCS1/count/ctrl_hi22_SOCS1_counts.tsv.gz
#>  [ reached 'max' / getOption("max.print") -- omitted 4138 rows ]

# Perform differential test
umi <- fisherUMI4C(umi)
results(umi)

#> GRanges object with 198 ranges and 6 metadata columns:
#>         seqnames            ranges strand |          id    pvalue odds_ratio
#>            <Rle>         <IRanges>  <Rle> | <character> <numeric>  <numeric>
#>     [1]    chr16 10849721-10854720      * |    region_1      <NA>       <NA>
#>     [2]    chr16 10854721-10859720      * |    region_2      <NA>       <NA>
#>     [3]    chr16 10859721-10864720      * |    region_3      <NA>       <NA>
#>     [4]    chr16 10864721-10869720      * |    region_4      <NA>       <NA>
#>     [5]    chr16 10869721-10874720      * |    region_5      <NA>       <NA>
#>     ...      ...               ...    ... .         ...       ...        ...
#>   [194]    chr16 11824721-11829720      * |  region_194      <NA>       <NA>
#>   [195]    chr16 11829721-11834720      * |  region_195      <NA>       <NA>
#>   [196]    chr16 11834721-11839720      * |  region_196      <NA>       <NA>
#>   [197]    chr16 11839721-11844720      * |  region_197      <NA>       <NA>
#>   [198]    chr16 11844721-11849720      * |  region_198      <NA>       <NA>
#>         log2_odds_ratio      padj      sign
#>               <numeric> <numeric> <logical>
#>     [1]            <NA>      <NA>      <NA>
#>     [2]            <NA>      <NA>      <NA>
#>     [3]            <NA>      <NA>      <NA>
#>     [4]            <NA>      <NA>      <NA>
#>     [5]            <NA>      <NA>      <NA>
#>     ...             ...       ...       ...
#>   [194]            <NA>      <NA>      <NA>
#>   [195]            <NA>      <NA>      <NA>
#>   [196]            <NA>      <NA>      <NA>
#>   [197]            <NA>      <NA>      <NA>
#>   [198]            <NA>      <NA>      <NA>
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths