UMI4Cats: A package for analyzing UMI-4C chromatin contact data

The UMI4Cats package provides functions for the pre-processing, analysis and visualization of UMI-4C chromatin contact data.

File preparation

There are two different functions that can be used to prepare the files for analyzing them with UMI4Cats:

demultiplexFastq. Demultiplex reads belonging to different viewpoints from a paired-end FastQ file.

digestGenome. Digest the reference genome of choice using a given restriction sequence.

Processing

The pre-processing functions are wrapped in the contactsUMI4C main function. This function will sequentially run the following steps:

prepUMI4C. Filter specific and high quality reads.

splitUMI4C. Split reads by the restriction sequence.

alignmentUMI4C. Align reads to the reference genome.

counterUMI4C. Apply UMI counting algorithm to quantify the interactions with the viewpoint.

The statistics from the samples analyzed with the contactsUMI4C function can be extracted and visualized with the function statsUMI4C.

Analysis

The analysis of UMI-4C data is wrapped in the construction of an object of UMI4C class by the creator function makeUMI4C. This function will group your samples according to the variable you provided in the grouping argument (default: "condition") and then normalize it to ref_umi4c.

The differential analysis can be performed with fisherUMI4C or differentialNbinomWaldTestUMI4C functions. Both will return a UMI4C object containing the results of the differential test. You can access these results with the method resultsUMI4C.

Visualization

An integrative plot showing the results stored inside the UMI4C object can be generated with the function plotUMI4C.