UMI4C alignment

Align splitted UMI-4C reads to a reference genome using Bowtie2.

alignmentUMI4C(
  wk_dir,
  pos_viewpoint,
  bowtie_index,
  threads = 1,
  filter_bp = 1e+07
)

Arguments

wk_dir
Working directory where to save the outputs generated by the UMI-4c analysis.

pos_viewpoint
GRanges object containing the genomic position of the viewpoint. It can be generated by `getViewpointCoordinates` function.

bowtie_index
Path and prefix of the bowtie index to use for the alignment.

threads
Number of threads to use in the analysis.

filter_bp
Integer indicating the bp upstream and downstream of the viewpoint to select for further analysis. Default: 10Mb.

Value

Creates a BAM file in wk_dir/align named "basename(fastq))_filtered.bam", containing the aligned filtered reads. The alignment log is also generated in wk_dir/logs named "umi4c_alignment_stats.txt".

Examples

path <- downloadUMI4CexampleData(use_sample=TRUE)

#> Will begin downloading datasets to /var/folders/6s/7ydjk_816j12lnyvsx292_qr0000gn/T//RtmpcWznVe/file2c9343df612c

#> untar: using cmd = ‘/usr/bin/tar -xf '/var/folders/6s/7ydjk_816j12lnyvsx292_qr0000gn/T//RtmpcWznVe/file2c9343df612c' -C './'’

#> Done writing UMI4C example files to ./

alignmentUMI4C(wk_dir=file.path(path, "SOCS1"),
               pos_viewpoint=GenomicRanges::GRanges(seqnames = 'chr16',
                IRanges(start=1134972, end=11349748)),
               bowtie_index=file.path(path, "ref_genome", "ucsc.hg19.chr16"))

#> 
#> [2020-03-06 22:14:19] Starting alignmentUMI4C using:
#>  > Work directory: ./UMI4Cats_data_sub/SOCS1 
#>  > Viewpoint position: chr16:1134972-11349748 
#>  > Reference genome: ./UMI4Cats_data_sub/ref_genome/ucsc.hg19.chr16 
#>  > Number of threads: 1

#> Warning: For argument `samOutput`, file exist:./UMI4Cats_data_sub/SOCS1/align/sub_ctrl_hi19_SOCS1_R1.sam. It will be overwrited

#> [2020-03-06 22:14:19] Finished sample sub_ctrl_hi19_SOCS1_R1.fq

#> Warning: For argument `samOutput`, file exist:./UMI4Cats_data_sub/SOCS1/align/sub_ctrl_hi19_SOCS1_R2.sam. It will be overwrited

#> [2020-03-06 22:14:20] Finished sample sub_ctrl_hi19_SOCS1_R2.fq