Call significant interactions

Test a set of query_regions for significant interactions with the viewpoint.

callInteractions(
  umi4c,
  design = ~condition,
  query_regions,
  padj_method = "fdr",
  zscore_threshold = 2,
  padj_threshold = 0.1,
  alpha = 20,
  penalty = 0.1
)

Arguments

umi4c
UMI4C object as generated by makeUMI4C or the UMI4C constructor.

design
A formula or matrix. The formula expresses how the counts for each fragment end depend on the variables in colData. See DESeqDataSet.

query_regions
GRanges object or data.frame containing the coordinates of the genomic regions you want to use to perform the analysis in specific genomic intervals. Default: NULL.

padj_method
The method to use for adjusting p-values, see p.adjust. Default: fdr.

zscore_threshold
Numeric indicating the z-score threshold to use to define significant differential contacts. Default: 2.

padj_threshold
Numeric indicating the adjusted p-value threshold to use to define significant differential contacts. Default: 0.1.

alpha
Approximate number of fragments desired for every basis function of the B-spline basis. floor((max(number of fragments)) / alpha) is passed to create.bspline.basis as nbasis argument. 4 is the minimum allowed value. Default: 20.

penalty
Amount of smoothing to be applied to the estimated functional parameter. Default: 0.1.

Value

GRangesList where each element is a UMI4C sample with the queried regions and their adjusted p-values and Z-scores.

Examples

data("ex_ciita_umi4c")
umi <- ex_ciita_umi4c
win_frags <- makeWindowFragments(umi, n_frags=8, sliding=1)

gr <- callInteractions(umi, ~condition, win_frags, padj_threshold = 0.01, zscore_threshold=2)

#> converting counts to integer mode

#> Warning: some variables in design formula are characters, converting to factors

#> 
#> [2021-05-11 13:05:20] Starting vstUMI4C
#>  > Samples of DDS object:
#>  ctrl_hi24_CIITA, ctrl_hi32_CIITA, cyt_hi24_CIITA, cyt_hi32_CIITA 

#> [2021-05-11 13:05:21] Finished vstUMI4C

#> 
#> [2021-05-11 13:05:21] Starting smoothMonotoneUMI4C using:
#>  > Samples of DDS object:
#>  ctrl_hi24_CIITA, ctrl_hi32_CIITA, cyt_hi24_CIITA, cyt_hi32_CIITA 
#>  > Alpha:
#>  20 
#>  > Penalty:
#>  0.1 

#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.3347      0.0076      5.2281      -0.6727
#> 1        0.3336      0.0058      5.379      -0.7141
#> 2        0.3335      0.0012      5.3969      -0.7191
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.1341      0.0056      5.2219      -0.6236
#> 1        0.1339      0.0015      5.2443      -0.63
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.3316      0.0221      5.2695      -0.6511
#> 1        0.3244      0.0176      5.6284      -0.7426
#> 2        0.3231      0.0037      5.6977      -0.7613
#> 3        0.3229      0.0028      5.7545      -0.7773
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.1633      0.0087      5.1198      -0.587
#> 1        0.162      0.0068      4.989      -0.5469
#> 2        0.1617      0.002      4.961      -0.5386

#> [2021-05-11 13:05:23] Finished smoothMonotoneUMI4C

inter <- getSignInteractions(gr)