UMI counting

Algorithm for counting and collapsing the number of UMIs supporting a specific ligation.

counterUMI4C(
  wk_dir,
  pos_viewpoint,
  res_enz,
  digested_genome,
  filter_bp = 1e+07
)

Arguments

wk_dir
Working directory where to save the outputs generated by the UMI-4c analysis.

pos_viewpoint
GRanges object containing the genomic position of the viewpoint.

res_enz
Character containing the restriction enzyme sequence.

digested_genome
Path for the digested genome file generated using the digestGenome function.

filter_bp
Integer indicating the bp upstream and downstream of the viewpoint to select for further analysis. Default: 10Mb.

Details

For collapsing different molecules into the same UMI, takes into account the ligation position and the number of UMI sequence mismatches.

Examples

if (FALSE) {
counterUMI4C(wk_dir="SOCS1",
            bait_seq="CCCAAATCGCCCAGACCAG",
            bait_pad="GCGCG",
            res_enz="GATC",
            digested_genome=hg19_dpnii,
            ref_gen="~/data/reference_genomes/hg19/hg19.fa")
}