Differential UMI4C contacts using DESeq2 Wald Test

Using a UMI4C object, infers the differences between conditions specified in design of the smooth monotone fitted values using a Wald Test from DESeq2 package.

differentialNbinomWaldTestUMI4C(
  umi4c,
  design = ~condition,
  normalized = TRUE,
  padj_method = "fdr",
  query_regions = NULL,
  padj_threshold = 0.05,
  penalty = 0.1,
  alpha = 20
)

Arguments

umi4c
UMI4C object as generated by makeUMI4C or the UMI4C constructor.

design
A formula or matrix. The formula expresses how the counts for each fragment end depend on the variables in colData. See DESeqDataSet.

normalized
Logical indicating if the function should return normalized or raw UMI counts. Default: TRUE.

padj_method
The method to use for adjusting p-values, see p.adjust. Default: fdr.

query_regions
GRanges object or data.frame containing the coordinates of the genomic regions you want to use to perform the analysis in specific genomic intervals. Default: NULL.

padj_threshold
Numeric indicating the adjusted p-value threshold to use to define significant differential contacts. Default: 0.05.

penalty
Amount of smoothing to be applied to the estimated functional parameter. Default: 0.1.

alpha
Approximate number of fragments desired for every basis function of the B-spline basis. floor((max(number of fragments)) / alpha) is passed to create.bspline.basis as nbasis argument. 4 is the minimum allowed value. Default: 20.

Value

UMI4C object with the DESeq2 Wald Test results.

Details

This function fits the signal trend of a variance stabilized count values using a symmetric monotone fit for the distance dependency. Then scales the raw counts across the samples to obtain normalized factors. Finally, it detects differences between conditions applying the DESeq2 Wald Test.

Examples

 files <- list.files(system.file("extdata", "CIITA", "count", package="UMI4Cats"),
                     pattern = "*_counts.tsv.gz",
                     full.names = TRUE
 )
# Create colData including all relevant information
colData <- data.frame(
  sampleID = gsub("_counts.tsv.gz", "", basename(files)),
  file = files,
  stringsAsFactors = FALSE
)

library(tidyr)
colData <- colData %>%
  separate(sampleID,
           into = c("condition", "replicate", "viewpoint"),
           remove = FALSE
  )
  
# Make UMI-4C object including grouping by condition
umi <- makeUMI4C(
  colData = colData,
  viewpoint_name = "CIITA",
  grouping = NULL,
  bait_expansion = 2e6
)

umi_wald <- differentialNbinomWaldTestUMI4C(umi4c=umi,
                                            design=~condition,
                                            alpha = 100)

#> Warning: some variables in design formula are characters, converting to factors

#> 
#> [2020-12-15 16:45:36] Starting vstUMI4C
#>  > Samples of DDS object:
#>  ctrl_hi24_CIITA, ctrl_hi32_CIITA, cyt_hi24_CIITA, cyt_hi32_CIITA 

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> [2020-12-15 16:45:37] Finished vstUMI4C

#> 
#> [2020-12-15 16:45:37] Starting smoothMonotoneUMI4C using:
#>  > Samples of DDS object:
#>  ctrl_hi24_CIITA, ctrl_hi32_CIITA, cyt_hi24_CIITA, cyt_hi32_CIITA 
#>  > Alpha:
#>  100 
#>  > Penalty:
#>  0.1 

#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.1605      0.002      1.3693      -0.1402
#> 1        0.1554      0.003      2.4147      -0.2588
#> 2        0.155      2e-04      2.4607      -0.2651
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.3019      0.0068      2.2821      -0.2848
#> 1        0.2944      0.006      2.9711      -0.3932
#> 2        0.2915      0.0028      3.3079      -0.4437
#> 3        0.2907      0.0016      3.5      -0.4759
#> 4        0.2905      8e-04      3.6235      -0.4962
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.1911      0.0027      1.529      -0.1656
#> 1        0.1853      0.0039      2.5313      -0.2908
#> 2        0.1844      4e-04      2.6208      -0.3035
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.2977      0.0051      2.1222      -0.2579
#> 1        0.2935      0.0043      2.5996      -0.3353
#> 2        0.2919      0.0019      2.8275      -0.3712
#> 3        0.2914      0.0012      2.9752      -0.3965
#> 4        0.2913      6e-04      3.0622      -0.4112

#> [2020-12-15 16:45:50] Finished smoothMonotoneUMI4C

#> 
#> [2020-12-15 16:45:50] Starting nbinomWaldTestUMI4C using:
#>  > Samples of DDS object:
#>  ctrl_hi24_CIITA, ctrl_hi32_CIITA, cyt_hi24_CIITA, cyt_hi32_CIITA 

#> found already estimated dispersions, replacing these

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.

#> final dispersion estimates

#> [2020-12-15 16:45:52] Finished sample nbinomWaldTestUMI4C