Differential UMI4C contacts using Fisher's Exact test

Using the UMIs inside query_regions performs Fisher's Exact test to calculate significant differences between contact intensities.

fisherUMI4C(
  umi4c,
  query_regions,
  resize = NULL,
  window_size = 5000,
  filter_low = 50,
  padj_method = "fdr",
  padj_threshold = 0.05
)

Arguments

umi4c
UMI4C object as generated by makeUMI4C or the UMI4C constructor.

query_regions
GenomicRanges object or data.frame containing the region coordinates used to perform the differential analysis.

resize
Width in base pairs for resizing the query_regions. Default: no resizing.

window_size
If query_regions are not defined, wil bin region in window_size bp and perform the analysis using this windows.

filter_low
Either the minimum median UMIs requiered to perform Fisher's Exact test or FALSE for performing the test in all windows.

padj_method
Method for adjusting p-values. See p.adjust for the different methods.

padj_threshold
Numeric indicating the adjusted p-value threshold to use to define significant differential contacts.

Value

Calculates statistical differences between UMI-4C experiments.

Details

This function calculates the overlap of fragment ends with either the provided query_regions or the binned region using window_size. The resulting number of UMIs in each query_region will be the sum of UMIs in all overlapping fragment ends. As a default, will filter out those regions whose median UMIs are lower than filter_low.

Finally, a contingency table for each query_reegions or window that passed the filter_low filter is created as follows:

query_region	region

Reference	n1	N1-n1

Condition	n2	N2-n2

and the Fisher's Exact test is performed. Obtained p-values are then converted to adjusted p-values using padj_method and the results list is added to the UMI4C object.

Examples

# Create example UMI4C-class object with helper function
umi <- makeUMI4Cexample()

# Perform differential test
umi <- fisherUMI4C(umi, filter_low = 20)
results(umi)

#> GRanges object with 198 ranges and 6 metadata columns:
#>         seqnames            ranges strand |          id    pvalue odds_ratio
#>            <Rle>         <IRanges>  <Rle> | <character> <numeric>  <numeric>
#>     [1]    chr16 10470996-10475995      * |    region_1        NA         NA
#>     [2]    chr16 10475996-10480995      * |    region_2        NA         NA
#>     [3]    chr16 10480996-10485995      * |    region_3        NA         NA
#>     [4]    chr16 10485996-10490995      * |    region_4        NA         NA
#>     [5]    chr16 10490996-10495995      * |    region_5        NA         NA
#>     ...      ...               ...    ... .         ...       ...        ...
#>   [194]    chr16 11445996-11450995      * |  region_194        NA         NA
#>   [195]    chr16 11450996-11455995      * |  region_195        NA         NA
#>   [196]    chr16 11455996-11460995      * |  region_196        NA         NA
#>   [197]    chr16 11460996-11465995      * |  region_197        NA         NA
#>   [198]    chr16 11465996-11470995      * |  region_198        NA         NA
#>         log2_odds_ratio      padj      sign
#>               <numeric> <numeric> <logical>
#>     [1]              NA        NA      <NA>
#>     [2]              NA        NA      <NA>
#>     [3]              NA        NA      <NA>
#>     [4]              NA        NA      <NA>
#>     [5]              NA        NA      <NA>
#>     ...             ...       ...       ...
#>   [194]              NA        NA      <NA>
#>   [195]              NA        NA      <NA>
#>   [196]              NA        NA      <NA>
#>   [197]              NA        NA      <NA>
#>   [198]              NA        NA      <NA>
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths