Get significant interactions from a GRangesList

Retrieves all significant interactions from the output of callInteractions.

getSignInteractions(gr_interactions)

Arguments

gr_interactions
GRangesList outputed by callInteractions.

Value

GRanges object with a list of significantly interacting regions.

Examples

data("ex_ciita_umi4c")
umi <- ex_ciita_umi4c
win_frags <- makeWindowFragments(umi, n_frags=8, sliding=1)

gr <- callInteractions(umi, ~condition, win_frags, padj_threshold = 0.01, zscore_threshold=2)

#> converting counts to integer mode

#> Warning: some variables in design formula are characters, converting to factors

#> 
#> [2021-10-28 14:23:01] Starting vstUMI4C
#>  > Samples of DDS object:
#>  ctrl_hi24_CIITA, ctrl_hi32_CIITA, cyt_hi24_CIITA, cyt_hi32_CIITA 

#> [2021-10-28 14:23:02] Finished vstUMI4C

#> 
#> [2021-10-28 14:23:02] Starting smoothMonotoneUMI4C using:
#>  > Samples of DDS object:
#>  ctrl_hi24_CIITA, ctrl_hi32_CIITA, cyt_hi24_CIITA, cyt_hi32_CIITA 
#>  > Alpha:
#>  20 
#>  > Penalty:
#>  0.1 

#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.3347      0.0076      5.2281      -0.6727
#> 1        0.3336      0.0058      5.379      -0.7141
#> 2        0.3335      0.0012      5.3969      -0.7191
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.1341      0.0056      5.2219      -0.6236
#> 1        0.1339      0.0015      5.2443      -0.63
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.3316      0.0221      5.2695      -0.6511
#> 1        0.3244      0.0176      5.6284      -0.7426
#> 2        0.3231      0.0037      5.6977      -0.7613
#> 3        0.3229      0.0028      5.7545      -0.7773
#> 
#>  Results 
#> 
#> Iter.   PENSSE   Grad Length Intercept   Slope
#> 0        0.1633      0.0087      5.1198      -0.587
#> 1        0.162      0.0068      4.989      -0.5469
#> 2        0.1617      0.002      4.961      -0.5386

#> [2021-10-28 14:23:03] Finished smoothMonotoneUMI4C

inter <- getSignInteractions(gr)