Get viewpoint coordinates

Finds the viewpoint coordinates for a given reference genome and sequence.

getViewpointCoordinates(
  bait_seq,
  bait_pad,
  res_enz,
  ref_gen,
  sel_seqname = NULL
)

Arguments

bait_seq
Character containing the bait primer sequence.

bait_pad
Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).

res_enz
Character containing the restriction enzyme sequence.

ref_gen
A BSgenome object of the reference genome.

sel_seqname
A character with the chromosome name to focus the search for the viewpoint sequence.

Value

Creates a GRanges object containing the genomic position of the viewpoint.

Examples

getViewpointCoordinates(
    bait_seq = "GGACAAGCTCCCTGCAACTCA",
    bait_pad = "GGACTTGCA",
    res_enz = "GATC",
    ref_gen = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
    sel_seqname = "chr16" # Look only in chr16
)

#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames            ranges strand
#>          <Rle>         <IRanges>  <Rle>
#>   [1]    chr16 10972515-10972548      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths