Split UMI4C reads

Split the prepared reads using the restrition enzyme information.

splitUMI4C(wk_dir, res_enz, cut_pos, numb_reads = 1e+11, min_flen = 20)

Arguments

wk_dir
Working directory where to save the outputs generated by the UMI-4c analysis.

res_enz
Character containing the restriction enzyme sequence.

cut_pos
Numeric indicating the nucleotide position where restriction enzyme cuts (zero-based) (for example, for DpnII is 0).

numb_reads
Number of lines from the FastQ file to load in each loop. If having memory size problems, change it to a smaller number. Default=10e10.

min_flen
Minimal fragment length to use for selecting the fragments.

Value

Creates a compressed FASTQ file in wk_dir/split named basename(fastq)).fq.gz, containing the splited reads based on the restriction enzyme used.

Examples

path <- downloadUMI4CexampleData(use_sample=TRUE)

#> Will begin downloading datasets to /var/folders/6s/7ydjk_816j12lnyvsx292_qr0000gn/T//RtmpcWznVe/file2c93193e4c50

#> untar: using cmd = ‘/usr/bin/tar -xf '/var/folders/6s/7ydjk_816j12lnyvsx292_qr0000gn/T//RtmpcWznVe/file2c93193e4c50' -C './'’

#> Done writing UMI4C example files to ./

splitUMI4C(wk_dir=file.path(path, "SOCS1"),
           res_enz="GATC",
           cut_pos=0)

#> 
#> [2020-03-06 22:23:02] Starting splitUMI4C using:
#>  > Work directory: ./UMI4Cats_data_sub/SOCS1 
#>  > Cut position: 0 
#>  > Restriction enzyme: GATC 
#>  > Number of reads loaded into memory: 1e+11

#> [2020-03-06 22:23:02] Finished sample sub_ctrl_hi19_SOCS1_R1.fq.gz

#> [2020-03-06 22:23:02] Finished sample sub_ctrl_hi19_SOCS1_R2.fq.gz

#> [[1]]
#> NULL
#>