DESeq2 Wald test for differential contacts

Using a UMI4C object, infers the differences between conditions specified in design using a Wald Test from DESeq2 package.

waldUMI4C(
  umi4c,
  query_regions = NULL,
  subset = "sum",
  design = ~condition,
  normalized = TRUE,
  padj_method = "fdr",
  padj_threshold = 0.05
)

Arguments

umi4c
UMI4C object as generated by makeUMI4C or the UMI4C constructor.

query_regions
GRanges object containing the coordinates of the genomic regions you want to use to perform the analysis in specific genomic intervals. Default: NULL.

subset
If query_regions are provided, how to subset the UMI4C object: "sum" for summing raw UMIs in fragments overlapping query_regions (default) or "overlap" for selecting overlapping fragments.

design
A formula or matrix. The formula expresses how the counts for each fragment end depend on the variables in colData. See DESeqDataSet.

normalized
Logical indicating if the function should return normalized or raw UMI counts. Default: TRUE.

padj_method
The method to use for adjusting p-values, see p.adjust. Default: fdr.

padj_threshold
Numeric indicating the adjusted p-value threshold to use to define significant differential contacts. Default: 0.05.

Value

UMI4C object with the DESeq2 Wald Test results, which can be accessed using resultsUMI4C.

Examples

data("ex_ciita_umi4c")

umi_dif <- waldUMI4C(ex_ciita_umi4c)

#> Warning: some variables in design formula are characters, converting to factors

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.

#> final dispersion estimates

#> fitting model and testing