UMI4C class methods

This page contains a summary of the different methods used to access the information contained inside the UMI4C-class object. See the details section for more information on the different accessors.

dgram(object)

dgram(object) <- value

groupsUMI4C(object, value)

groupsUMI4C(object) <- value

bait(object)

trend(object)

resultsUMI4C(object, format = "GRanges", counts = TRUE, ordered = FALSE)

# S4 method for UMI4C
dgram(object)

# S4 method for UMI4C
dgram(object) <- value

# S4 method for UMI4C
groupsUMI4C(object)

# S4 method for UMI4C
groupsUMI4C(object) <- value

# S4 method for UMI4C
bait(object)

# S4 method for UMI4C
trend(object)

# S4 method for UMI4C
resultsUMI4C(object, format = "GRanges", counts = FALSE, ordered = FALSE)

Arguments

object
a UMI4C-class object.

value
Alternative list of dgrams to replace the current slot.

format
Either "GRanges" (default) or "data.frame", indicating the format output of the results.

counts
Logical indicating whether counts for the different region should be provided. Default: FALSE.

ordered
Logical indicating whether to sort output by significance (adjusted p-value). Default: FALSE.

Value

There are several accessors to easily retrive information from a UMI4C-class object:

dgram: Returns a named list with the output domainograms for each sample.

bait: Returns a GRanges object with the position of the bait.

trend: Returns a data.frame in long format with the values of the adapative smoothen trend.

resultsUMI4C: Returns a GRanges or data.frame with the results of the differential analysis.

See also

Examples

# Access the different information inside the UMI4C object
data("ex_ciita_umi4c")
ex_ciita_umi4c <- addGrouping(ex_ciita_umi4c, grouping="condition")

dgram(ex_ciita_umi4c)

#> List of length 4
#> names(4): ctrl_hi24_CIITA ctrl_hi32_CIITA cyt_hi24_CIITA cyt_hi32_CIITA

bait(ex_ciita_umi4c)

#> GRanges object with 1 range and 1 metadata column:
#>                   seqnames            ranges strand |        name
#>                      <Rle>         <IRanges>  <Rle> | <character>
#>   ctrl_hi24_CIITA    chr16 10970996-10972544      * |       CIITA
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

head(trend(ex_ciita_umi4c))

#>   geo_coord     trend         sd scale id_contact          sample sampleID
#> 1  10834435 0.1477273 0.08194435    44  frag_2481 ctrl_hi24_CIITA     <NA>
#> 2  10834681 0.1511628 0.08385003    43  frag_2482 ctrl_hi24_CIITA     <NA>
#> 3  10834926 0.1547619 0.08584646    42  frag_2483 ctrl_hi24_CIITA     <NA>
#> 4  10835168 0.1585366 0.08794028    41  frag_2484 ctrl_hi24_CIITA     <NA>
#> 5  10835417 0.1625000 0.09013878    40  frag_2485 ctrl_hi24_CIITA     <NA>
#> 6  10835672 0.1666667 0.09245003    39  frag_2486 ctrl_hi24_CIITA     <NA>
#>   replicate viewpoint file
#> 1      <NA>      <NA> <NA>
#> 2      <NA>      <NA> <NA>
#> 3      <NA>      <NA> <NA>
#> 4      <NA>      <NA> <NA>
#> 5      <NA>      <NA> <NA>
#> 6      <NA>      <NA> <NA>

# Perform differential test
enh <- GRanges(c("chr16:10925006-10928900", "chr16:11102721-11103700"))
umi_dif <- fisherUMI4C(ex_ciita_umi4c, query_regions = enh, 
                       filter_low = 20, resize = 5e3)
resultsUMI4C(umi_dif)

#> GRanges object with 2 ranges and 6 metadata columns:
#>       seqnames            ranges strand |          id      pvalue odds_ratio
#>          <Rle>         <IRanges>  <Rle> | <character>   <numeric>  <numeric>
#>   [1]    chr16 10924453-10929452      * |    region_1 1.28944e-05    2.55569
#>   [2]    chr16 11100711-11105710      * |    region_2          NA         NA
#>       log2_odds_ratio        padj      sign
#>             <numeric>   <numeric> <logical>
#>   [1]         1.35371 1.28944e-05      TRUE
#>   [2]              NA          NA      <NA>
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths