Split UMI4C reads

Split the prepared reads using the restrition enzyme information.

splitUMI4C(wk_dir, res_enz, cut_pos, numb_reads = 1e+11, min_flen = 20)

Arguments

wk_dir
Working directory where to save the outputs generated by the UMI-4c analysis.

res_enz
Character containing the restriction enzyme sequence.

cut_pos
Numeric indicating the nucleotide position where restriction enzyme cuts (zero-based) (for example, for DpnII is 0).

numb_reads
Number of lines from the FastQ file to load in each loop. If having memory size problems, change it to a smaller number. Default=10e10.

min_flen
Minimal fragment length to use for selecting the fragments. Default=20

Value

Creates a compressed FASTQ file in wk_dir/split named basename(fastq)).fq.gz, containing the split reads based on the restriction enzyme used.

Examples

path <- downloadUMI4CexampleData(reduced = TRUE)

#> untar: using cmd = ‘/bin/tar -xf '/tmp/Rtmp9jKRrg/BiocFileCache/276f7df1385d_UMI4Cats_data_reduced.tar.gz' -C '/tmp/Rtmp9jKRrg'’

splitUMI4C(
    wk_dir = file.path(path, "CIITA"),
    res_enz = "GATC",
    cut_pos = 0
)

#> 
#> [2020-07-30 18:22:40] Starting splitUMI4C using:
#>  > Work directory: /tmp/Rtmp9jKRrg/UMI4Cats_data_reduced/CIITA 
#>  > Cut position: 0 
#>  > Restriction enzyme: GATC 
#>  > Number of reads loaded into memory: 1e+11

#> [2020-07-30 18:22:40] Finished sample ctrl_hi24_CIITA_R1.fq.gz

#> [2020-07-30 18:22:41] Finished sample ctrl_hi24_CIITA_R2.fq.gz

#> [[1]]
#> NULL
#> 

unlink(path, recursive=TRUE)